Lee Elliott:
while I was researching the MiG-15bis I found a reference to the
approach speed as being 93 kts and although I thought this was
far too low I went with it until I found a better source of info
- well I have now.  I found a good personal account of flying
the MiG-15UT (the dual seat trainer version) and in this the
pilot says that he comes over the fence at 125 kts - a much more
realistic speed.

Anyway, I've incorporated this value into the YASim approach
section and considering that this is quite a major change I
thought I'd get an update to you.

The 3d model has been updated as well but is still un-textured.

The YASim config still needs quite a bit of work as there are a
few quirks I want to sort out.  One of them is that if I use low
altitude datum (0 ft) in the cruise section, which should define
the max speed at that alt, it can actually exceed it by about 15
kt.  On the other hand, if I use high altitude datum it can't
reach the max speed at 0 ft - it's about 100 kt too slow.

Therefore I've plumped for the low alt datum and while it's a
struggle to achieve it's ceiling it can just about sustain
flight at 50,000 ft, at around the right speed.

A few other files have been changed too.